# perf(trie-router): improve performance and reduce file size