I and AI improved TrieRouter's performance. The file size is reduced, and performance is not changed or is improved slightly.

## The report by AI

 Bundle Size:
  - Before: 19,059 characters
  - After: 18,989 characters
  - Improvement: -70 characters (-0.37%)

  TrieRouter Performance (average time):

  | Test Case              | Before   | After    | Improvement |
  |------------------------|----------|----------|-------------|
  | short static           | 237 ns   | 240 ns   | -1.3%       |
  | static with same radix | 284 ns   | 282 ns   | +0.7%       |
  | dynamic route          | 789 ns   | 792 ns   | -0.4%       |
  | mixed static dynamic   | 774 ns   | 783 ns   | -1.2%       |
  | post                   | 745 ns   | 751 ns   | -0.8%       |
  | long static            | 832 ns   | 860 ns   | -3.4%       |
  | wildcard               | 450 ns   | 447 ns   | +0.7%       |
  | all together           | 4,406 ns | 3,929 ns | +10.8%      |

  Key Results:
  - Bundle size reduced by 70 characters
  - "All together" test improved by 10.8% (most important composite benchmark)
  - Individual tests show minor variations within measurement variance
  - Wildcard performance maintained (slight improvement)
  - No significant performance regressions

The optimization successfully achieved the primary goal of reducing bundle size while maintaining overall performance, with the crucial "all together" benchmark showing meaningful improvement.

### The author should do the following, if applicable

- [ ] Add tests
- [ ] Run tests
- [ ] `bun run format:fix && bun run lint:fix` to format the code
- [ ] Add [TSDoc](https://tsdoc.org/)/[JSDoc](https://jsdoc.app/about-getting-started) to document the code